backend/requirements.txt
jsonschema~=4.19.0
pandas~=2.0.3
openpyxl~=3.1.2
numpy~=1.25.2
sweref99~=0.2
uvicorn[standard]~=0.23.2
fastapi~=0.101.1
pydantic-settings~=2.0.3
python-multipart~=0.0.6
asyncpg~=0.28.0
sqlalchemy[asyncio]~=2.0.21
alembic~=1.11.3
GeoAlchemy2~=0.14.1
types-jsonschema==4.17.0.10
types-pytz==2023.3.0.1
aiohttp~=3.8.5
celery[redis]~=5.3.3
python-keycloak~=3.3.0
